Updated NutritionPanel.java
Added the ability to swap between different panels using a CardLayout for the center most panel. This will be updated more to be specifically what we want in the future but for now, this works to be able to swap panels for the center using the Dropdown on the top.